worker: address Task 1 review (static httpd buffer, dedup bin-width const)
diag_dcfine_get's 4096-byte body buffer moves from stack to static (matches
the existing static char body[1024] pattern in the same file; the httpd
server runs a single worker task so this is safe). The rounded
1221 Hz reporting bin width is now WORKER_DCFINE_BIN_HZ_REPORT /
WORKER_DCFINE_BIN0_HZ_REPORT in worker_dcfine.h instead of duplicated
literals, with a comment distinguishing it from the exact bucketing
value used in worker_dcfine_index(). Also drops a redundant (float)
cast on an already-float field in worker_core1.c.

Smoke-verified: RAW=PASS(matched=58) FRAME=PASS CORPUS=FAIL(pre-existing:live-RFI-beats-DC-tone) REAL=no-verdict(8-stripe>cap,detecting-normally); dcmask verified-disabled + RAW-pass => regression-free

+++ p4-usb-host/main/worker_core1.c
@@ -1258,7 +1258,7 @@ void worker_core1_push_burst(const detected_burst_t *burst)
     if (!s_pq_lock) return;
     s_bursts_queued++;
     hist_freq_record(burst->rel_freq_hz);                        // band occupancy of ALL detections
-    hist_dcfine_record((float)burst->rel_freq_hz);               // fine near-DC diagnostic
+    hist_dcfine_record(burst->rel_freq_hz);                      // fine near-DC diagnostic
     hist_snr_record_into(s_hist_snr_pushed, burst->peak_snr_db); // P1.5c: SNR of ALL detections
     hist_duration_record_pushed(burst->length_samples);          // P1.5c: duration class of ALL detections
